reporter: moved request handlers to specific package
